CI: fix rebuilding peepmatic peephole optimizers (#2311)
The test that triggers the rebuild of the peephole optimizers is in the
`cranelift-codegen` crate, not the umbrella cranelift crate. This was previously
successfully running zero tests, and then successfully reporting no `git diff`
because no peephole optimizers were ever rebuilt.

This change fixes it so that we run the correct test that triggers the
rebuilding of the peephole optimizers.